main currency bug fixed + change products image dimensions

+++ app/Http/Controllers/Admin/CurrencyController.php
@@ -121,8 +121,11 @@ class CurrencyController extends Controller
      */
     public function destroy(Request $request, Currency $currency): RedirectResponse
     {
+        if($currency->code == 'USD' || $currency->symbol == '$'){
+            return back()->with('error','Main currency can\'t be deleted');
+        }
         $this->currencyRepository->delete($currency);
-        $request->session()->put('appcurrency',getMainCurrency()->code);
+        $request->session()->put('appcurrency',getMainCurrency()->code ?? 'USD');
         $request->session()->flash('success', __($this->resource.'.'.$this->resource.'_deleted_successfully'));
         return redirect()->route('admin.currencies.index');
